% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BlosumVec}
\alias{BlosumVec}
\title{Amino acid pairs with non-negative BLOSUM62 score}
\format{
An object of class \code{character} of length 112.
}
\usage{
data(BlosumVec)
}
\description{
\code{"BlosumVec"} is a vector whose elements represent amino acid pairs in the one-letter code that have a
BLOSUM62 score >= 0. This information is used by GLIPH2 to cluster only global similarities whose different
amino acids are interpreted as interchangeable according to the BLOSUM62 matrix.
}
\keyword{datasets}
