% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_motifs}
\alias{find_motifs}
\title{Finding motifs in sequences}
\usage{
find_motifs(seqs, q = 2:4, kmer_mindepth = NULL, discontinuous = FALSE)
}
\arguments{
\item{seqs}{character vector. This vector must contain the sequences whose motifs are to be identified and quantified.}

\item{q}{accepts a numeric vector of motif lengths
you want to find. By default it searches for
motifs of size 2, 3 and 4.}

\item{kmer_mindepth}{numeric. By default 3. Minimum observations of kmer for it to be evaluated. This is
the minimum number of times a kmer should be observed in
the sample set in order for it to be considered for being returned.}

\item{discontinuous}{logical. By default \code{FALSE}. Determines whether discontinuous motifs are to be considered.}
}
\value{
\code{find_motifs} returns a data frame with two columns.
The first column contains the motifs and the second column the frequency of the motifs.
}
\description{
Finding motifs in sequences. This function searches and simultaneously counts continuous and discontinuous motifs in a sequence vector.
It is used in the \code{turbo_gliph} function to identify local similarities.
}
\examples{
utils::data("gliph_input_data")
sample_seqs <- base::as.character(gliph_input_data$CDR3b)
res <- find_motifs(seqs = sample_seqs)

}
