% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRandomSubsample.R
\name{getRandomSubsample}
\alias{getRandomSubsample}
\title{Generating a random subsample}
\usage{
getRandomSubsample(
  cdr3_len_stratify = FALSE,
  vgene_stratify = FALSE,
  refseqs_motif_region,
  motif_region,
  motif_lengths_list,
  ref_motif_lengths_id_list,
  motif_region_vgenes_list,
  ref_motif_vgenes_id_list,
  ref_lengths_vgenes_list,
  lengths_vgenes_list
)
}
\arguments{
\item{cdr3_len_stratify}{logical. By default \code{FALSE}.
Specifies whether the distribution of the cdr3 lengths in the sample should be retained during repeat random sampling.}

\item{vgene_stratify}{logical. By default \code{FALSE}.
Specifies whether the distribution of V-genes in the sample should be retained during repeat random sampling.}

\item{refseqs_motif_region}{character vector. Contains the motif regions of reference sequences.}

\item{motif_region}{character vector. Contains the motif regions of sample sequences.}

\item{motif_lengths_list}{list. Required if \code{cdr3_len_stratify = TRUE}.
The elements are named after the different cdr3 lengths in the \code{motif_region} vector
and contain the frequency of occurrence of the corresponding cdr3 length in the \code{motif_region} vector.}

\item{ref_motif_lengths_id_list}{list. Required if \code{cdr3_len_stratify = TRUE}.
The elements are named after the different cdr3 lengths in the \code{motif_region} vector
and contain the indices of sequences in the \code{refseqs_motif_region} vector with the corresponding cdr3 length.}

\item{motif_region_vgenes_list}{list. Required if \code{vgene_stratify = TRUE}.
The elements are named after the different V-genes of the sequences in the \code{motif_region} vector
and contain the frequency of occurrence of the corresponding V-genes of the sequences in the \code{motif_region} vector.}

\item{ref_motif_vgenes_id_list}{list. Required if \code{vgene_stratify = TRUE}.
The elements are named after the different V-genes of the sequences in the \code{motif_region} vector
and contain the indices of sequences in the \code{refseqs_motif_region} vector with the corresponding V-gene.}

\item{ref_lengths_vgenes_list}{list. Required if \code{cdr3_len_stratify = TRUE} and \code{vgene_stratify = TRUE}.
The elements are lists itself and are named after the different cdr3 lengths in the \code{motif_region} vector.
The elements of any list are named after the different V-genes of the sequences in the \code{motif_region} vector
and contain the frequency of simultaneous occurrence of the corresponding cdr3 length and V-gene of the sequences in the \code{refseqs_motif_region} vector.}

\item{lengths_vgenes_list}{list. Required if \code{cdr3_len_stratify = TRUE} and \code{vgene_stratify = TRUE}.
The elements are lists itself and are named after the different cdr3 lengths in the \code{motif_region} vector.
The elements of any list are named after the different V-genes of the sequences in the \code{motif_region} vector
and contain the frequency of simultaneous occurrence of the corresponding cdr3 length and V-gene of the sequences in the \code{motif_region} vector.}
}
\value{
\code{getRandomSubsample} returns a character vector containing a subsample of \code{refseqs_motif_region} with the same size as \code{motif_region}
}
\description{
Generating a random subsample. This function is used by the \code{turbo_gliph} function to generate random subsamples of naive reference
sequences with a similar size as the sample size for random repeat sampling.
If specified the function tries to maintain the distribution of cdr3 lengths and/or V-gene usage of the whole sample in the subsample.
}
