% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gliph_input_data}
\alias{gliph_input_data}
\title{GLIPH input data}
\format{
An object of class \code{data.frame} with 2067 rows and 11 columns.
}
\source{
Glanville, Jacob, et al.
"Identifying specificity groups in the T cell receptor repertoire." Nature 547.7661 (2017): 94.
}
\usage{
data(gliph_input_data)
}
\description{
gliph_input_data, consists of around 2000 unique TCRs of known specificity
}
\keyword{datasets}
