% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gTRB}
\alias{gTRB}
\title{Germline TCR-beta CDR3 fragments}
\format{
An object of class \code{list} of length 3.
}
\source{
Lefranc, M.-P.
IMGT, the international ImMunoGeneTics database.
Nucl. Acids Res., 29(1):207-209 (2001). DOI:10.1093/nar/29.1.207. PMID:11125093.
}
\usage{
data(gTRB)
}
\description{
\code{"gTRB"} is a list containing three data frames for fragments of germline encoded V genes (\code{"gTRV"}),
D genes (\code{"gTRD"}) and J genes (\code{"gTRJ"}) that might be located in the CDR3 region. These fragments are
used by GLIPH2 to determine germline-encoded sequence segments.
}
\keyword{datasets}
