% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gliph_output.R
\name{load_gliph_output}
\alias{load_gliph_output}
\title{Loading turboGLIPH and GLIPH2 output}
\usage{
load_gliph_output(result_folder = "")
}
\arguments{
\item{result_folder}{character. Must be specified. Path to the folder in which the output files of the  gliph function are stored.}
}
\value{
\code{load_gliph_output} returns a list equivalent to either the \code{turboGliph} or \code{gliph2} function (see documentation
of the functions for more information).
}
\description{
Loading turboGLIPH and GLIPH2 output. This function loads the output of the GLIPH functions stored in txt files.
}
