% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_check.R
\name{parameter_check}
\alias{parameter_check}
\title{Check turboGliph Input parameters}
\usage{
parameter_check(...)
}
\arguments{
\item{...}{turboGliph parameters}
}
\value{
Returns checked / corrected parameters
}
\description{
Check turboGliph Input parameters
}
