% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_cluster_sizes}
\alias{ref_cluster_sizes}
\title{Cluster size probabilities in naive reference repertoire}
\format{
An object of class \code{list} of length 2.
}
\source{
Glanville, Jacob, et al.
"Identifying specificity groups in the T cell receptor repertoire." Nature 547.7661 (2017): 94.
}
\usage{
data(ref_cluster_sizes)
}
\description{
\code{"ref_cluster_sizes"} is a list containing two matrices with probabilities for cluster sizes in a naive reference
repertoire. The element named \code{"original"} contains the probabilities used in the original GLIPH algorithm. Here,
the same probabilities are used for all sample sizes.
The element names \code{"simulated"} contains probabilities determined by us. Since the distribution of cluster sizes
depend on the sample size, we estimated with GLIPH the probabilities for different sample sizes in a 500-step simulation
using different numbers of random sequences from the reference database (125, 250, 500, 1000, 2000, 4000, 6000, 8000,
10000). For scoring, the probabilities from the dataset corresponding the closest to the sample size are used.
}
\keyword{datasets}
