% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_list}
\alias{reference_list}
\title{Naive reference repertoires}
\format{
An object of class \code{list} of length 1.
}
\source{
Glanville, Jacob, et al.
"Identifying specificity groups in the T cell receptor repertoire." Nature 547.7661 (2017): 94.
}
\usage{
data(reference_list)
}
\description{
A list of naive reference repertoires. Each element of the list consists of
another list with a data frame containing the sequences and V-gene information, a data frame containing the V-gene
distribution and a data frame containing the CDR3 length distribution. 
The following database is available:
\itemize{
\item "gliph_reference": Reference database of 162,165 CDR3b sequences of naive human CD4+ or CD8+ T cells of two individuals used for the GLIPH paper.
}
Additional reference databases were provided for download by the developers of GLIPH2 in the web tool (http://50.255.35.37:8080/tools).
}
\keyword{datasets}
